refactor(ui): relocate compound organism components
