+++ src/app/(main)/_components/AnnouncementBar.test.tsx
