+++ src/app/(main)/_components/AnnouncementBar.tsx
