+++ src/ui/components/organisms/ClipboardButton.tsx
@@ -8,7 +8,7 @@ import { cn } from "@/core/utils";
 import { type VariantProps } from "class-variance-authority";
 
 /**
- * 클립보드 복사 버튼 UI (Molecule)
+ * 클립보드 복사 버튼 UI (Organism)
  * 복사 상태(isCopied)와 복사 실행 함수(onCopy)를 외부로부터 주입받는 순수 컴포넌트입니다.
  */
 export type ClipboardButtonProps = ComponentProps<"button"> &
